PLMDocAnalyzer/plm_digitizer/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
celery>=5.4.0
redis>=5.0.3
sqlalchemy>=2.0.29
pydantic>=2.6.4
pdfplumber>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.10
python-docx>=1.1.0
openpyxl>=3.1.2
openai>=1.25.0
cryptography>=43.0.0
aiofiles>=23.2.1
python-multipart>=0.0.9
httpx>=0.27.0
pillow>=11.0.0
websockets>=12.0
